Added Blog
I know this seams like a silly idea, but it has purpose.

I want to be able to update the landing page content without checking
content into the repository. So as to prevent polluting the commit log
with messages like, 'fixed grammar', and to make it easier for a non
technical editor to update content.